chore: Temporarily migrate tests
