+++ tests/_Original/Http/Resolvers/SubdomainResolverTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Http\Resolvers;
+namespace Sprout\Tests\_Original\Http\Resolvers;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -14,8 +14,9 @@ use Sprout\Attributes\CurrentTenant;
 use Sprout\Contracts\Tenant;
 use Workbench\App\Models\TenantModel;
 use function Sprout\resolver;
+use function Sprout\sprout;
 
-class PathResolverTest extends TestCase
+class SubdomainResolverTest extends TestCase
 {
     use WithWorkbench, RefreshDatabase;
 
@@ -25,53 +26,56 @@ class PathResolverTest extends TestCase
     {
         tap($app['config'], static function (Repository $config) {
             $config->set('multitenancy.providers.tenants.model', TenantModel::class);
-            $config->set('multitenancy.defaults.resolver', 'path');
+            $config->set('multitenancy.resolvers.subdomain.domain', 'localhost');
+            $config->set('session.driver', 'database');
         });
     }
 
     protected function withManualParameterName($app): void
     {
         tap($app['config'], static function (Repository $config) {
-            $config->set('multitenancy.resolvers.path.parameter', 'custom-parameter');
+            $config->set('multitenancy.resolvers.subdomain.parameter', 'custom-parameter');
         });
     }
 
     protected function withParameterPatternName($app): void
     {
         tap($app['config'], static function (Repository $config) {
-            $config->set('multitenancy.resolvers.path.pattern', '.*');
+            $config->set('multitenancy.resolvers.subdomain.pattern', '.*');
         });
     }
 
     protected function withoutManualParameterName($app): void
     {
         tap($app['config'], static function (Repository $config) {
-            $config->set('multitenancy.resolvers.path.parameter', null);
+            $config->set('multitenancy.resolvers.subdomain.parameter', null);
         });
     }
 
     protected function withoutParameterPattern($app): void
     {
         tap($app['config'], static function (Repository $config) {
-            $config->set('multitenancy.resolvers.path.pattern', null);
+            $config->set('multitenancy.resolvers.subdomain.pattern', null);
         });
     }
 
-    protected function defineRoutes($router): void
+    protected function defineRoutes($router)
     {
-        $router->get('/', function () {
-            return 'no';
-        });
+        $router->middleware('web')->group(function (Router $router) {
+            $router->get('/', function () {
+                return 'no';
+            });
 
-        $router->tenanted(function (Router $router) {
-            $router->get('/path-route', function (#[CurrentTenant] Tenant $tenant) {
+            $router->tenanted(function (Router $router) {
+                $router->get('/subdomain-route', function (#[CurrentTenant] Tenant $tenant) {
+                    return $tenant->getTenantKey();
+                })->name('subdomain.route');
+            }, 'subdomain', 'tenants');
+
+            $router->get('/subdomain-request', function (#[CurrentTenant] Tenant $tenant) {
                 return $tenant->getTenantKey();
-            })->name('path.route');
-        }, 'path', 'tenants');
-
-        $router->get('/{identifier}/path-request', function (#[CurrentTenant] Tenant $tenant) {
-            return $tenant->getTenantKey();
-        })->middleware('sprout.tenanted')->name('path.request');
+            })->middleware('sprout.tenanted')->name('subdomain.request');
+        });
     }
 
     #[Test]
@@ -79,10 +83,21 @@ class PathResolverTest extends TestCase
     {
         $tenant = TenantModel::factory()->createOne();
 
-        $result = $this->get(route('path.route', ['tenants_path' => $tenant->getTenantIdentifier()]));
+        $result = $this->get(route('subdomain.route', ['tenants_subdomain' => $tenant->getTenantIdentifier()]));
 
         $result->assertOk();
         $result->assertContent((string)$tenant->getTenantKey());
+        $this->assertTrue(sprout()->withinContext());
+    }
+
+    #[Test]
+    public function canAccessNonTenantedRoutesSuccessfully(): void
+    {
+        $result = $this->get('/');
+
+        $result->assertOk();
+        $result->assertContent('no');
+        $this->assertFalse(sprout()->withinContext());
     }
 
     #[Test]
@@ -90,7 +105,7 @@ class PathResolverTest extends TestCase
     {
         $tenant = TenantModel::factory()->createOne();
 
-        $result = $this->get('/' . $tenant->getTenantIdentifier() . '/path-request');
+        $result = $this->get('http://' . $tenant->getTenantIdentifier() . '.localhost/subdomain-request');
 
         $result->assertOk();
         $result->assertContent((string)$tenant->getTenantKey());
@@ -99,7 +114,7 @@ class PathResolverTest extends TestCase
     #[Test]
     public function throwsExceptionForInvalidTenantWithParameter(): void
     {
-        $result = $this->get(route('path.route', ['tenants_path' => 'i-am-not-real']));
+        $result = $this->get(route('subdomain.route', ['tenants_subdomain' => 'i-am-not-real']));
 
         $result->assertInternalServerError();
     }
@@ -107,7 +122,15 @@ class PathResolverTest extends TestCase
     #[Test]
     public function throwsExceptionForInvalidTenantWithoutParameter(): void
     {
-        $result = $this->get('/i-am-not-real/path-request');
+        $result = $this->get('http://i-am-not-real.localhost/subdomain-request');
+
+        $result->assertInternalServerError();
+    }
+
+    #[Test]
+    public function throwsExceptionForInvalidTenantWithJustMultitenantedDomain(): void
+    {
+        $result = $this->get('http://localhost/subdomain-request');
 
         $result->assertInternalServerError();
     }
@@ -116,7 +139,7 @@ class PathResolverTest extends TestCase
     public function hasNoParameterPatternByDefault(): void
     {
         /** @var \Sprout\Http\Resolvers\SubdomainIdentityResolver $resolver */
-        $resolver = resolver('path');
+        $resolver = resolver('subdomain');
 
         $this->assertNull($resolver->getPattern());
     }
@@ -125,7 +148,7 @@ class PathResolverTest extends TestCase
     public function hasDefaultParameterNameByDefault(): void
     {
         /** @var \Sprout\Http\Resolvers\SubdomainIdentityResolver $resolver */
-        $resolver = resolver('path');
+        $resolver = resolver('subdomain');
 
         $this->assertSame('{tenancy}_{resolver}', $resolver->getParameter());
     }
@@ -133,8 +156,8 @@ class PathResolverTest extends TestCase
     #[Test, DefineEnvironment('withManualParameterName')]
     public function allowsForCustomParameterName(): void
     {
-        /** @var \Sprout\Http\Resolvers\PathIdentityResolver $resolver */
-        $resolver = resolver('path');
+        /** @var \Sprout\Http\Resolvers\SubdomainIdentityResolver $resolver */
+        $resolver = resolver('subdomain');
 
         $this->assertSame('custom-parameter', $resolver->getParameter());
     }
@@ -142,8 +165,8 @@ class PathResolverTest extends TestCase
     #[Test, DefineEnvironment('withParameterPatternName')]
     public function allowsForCustomParameterPattern(): void
     {
-        /** @var \Sprout\Http\Resolvers\PathIdentityResolver $resolver */
-        $resolver = resolver('path');
+        /** @var \Sprout\Http\Resolvers\SubdomainIdentityResolver $resolver */
+        $resolver = resolver('subdomain');
 
         $this->assertSame('.*', $resolver->getPattern());
     }
