+++ tests/_Original/Listeners/SetCurrentTenantForJobTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests\Listeners;
+namespace Sprout\Tests\_Original\Listeners;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -15,10 +15,8 @@ use Sprout\Managers\TenancyManager;
 use Sprout\Overrides\AuthOverride;
 use Sprout\Overrides\CacheOverride;
 use Sprout\Overrides\CookieOverride;
-use Sprout\Overrides\JobOverride;
 use Sprout\Overrides\SessionOverride;
 use Sprout\Overrides\StorageOverride;
-use Sprout\TenancyOptions;
 use Workbench\App\Jobs\TestTenantJob;
 use Workbench\App\Models\TenantModel;
