+++ tests/_Original/Overrides/CookieOverrideTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests\Overrides;
+namespace Sprout\Tests\_Original\Overrides;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -16,12 +16,9 @@ use Sprout\Attributes\CurrentTenant;
 use Sprout\Contracts\Tenant;
 use Sprout\Overrides\AuthOverride;
 use Sprout\Overrides\CacheOverride;
-use Sprout\Overrides\CookieOverride;
 use Sprout\Overrides\JobOverride;
-use Sprout\Overrides\SessionOverride;
 use Sprout\Overrides\StorageOverride;
 use Workbench\App\Models\TenantModel;
-use function Sprout\sprout;
 
 #[Group('services'), Group('cookies')]
 class CookieOverrideTest extends TestCase
