+++ tests/_Original/Overrides/StorageOverrideTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests\Overrides;
+namespace Sprout\Tests\_Original\Overrides;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
