+++ tests/_Original/Providers/DatabaseTenantProviderTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests\Providers;
+namespace Sprout\Tests\_Original\Providers;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
