+++ tests/_Original/SproutTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Sprout\Tests;
+namespace Sprout\Tests\_Original;
 
 use Illuminate\Config\Repository;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -10,7 +10,6 @@ use Orchestra\Testbench\TestCase;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Test;
 use Sprout\Sprout;
-use Sprout\Support\ResolutionHook;
 use Workbench\App\Models\TenantModel;
 
 #[Group('core')]
